[10] Admin Panel V0.0.1 + more...
Not Much but added the template and did some other changes.
All will be reconsidered and maybe removed or changed.
More than 2k files have been added and at this stage the CMS is a bit
unstable. More will be removed and added in the future! I Hope you will
like the install feature and the Admin Template.